requirements.txt
dnspython
tabulate
pyyaml
colorama